+++ CHANGELOG.md
@@ -12,7 +12,8 @@ options. The project adheres to semantic versioning.
 * Fix duplicate default options when there is a short and long option.
 * Add `CXXOPTS_NO_EXCEPTIONS` to disable exceptions.
 * Fix char parsing for space and check for length.
-* Change argument type in `Options::parse` from `char**` to `const char**`.
+* Add `const char**` type to `Options::parse`.
+* Remove reference to argv in `Options::parse`.
 
 ## 2.2